+++ crates/sui-core/src/post_consensus_tx_reorder.rs
@@ -10,18 +10,17 @@ use sui_types::{
 pub struct PostConsensusTxReorder {}
 
 impl PostConsensusTxReorder {
-    // TODO: Remove the above versions and rename these without _v2
-    pub fn reorder_v2(
+    pub fn reorder(
         transactions: &mut [VerifiedExecutableTransaction],
         kind: ConsensusTransactionOrdering,
     ) {
         match kind {
-            ConsensusTransactionOrdering::ByGasPrice => Self::order_by_gas_price_v2(transactions),
+            ConsensusTransactionOrdering::ByGasPrice => Self::order_by_gas_price(transactions),
             ConsensusTransactionOrdering::None => (),
         }
     }
 
-    fn order_by_gas_price_v2(transactions: &mut [VerifiedExecutableTransaction]) {
+    fn order_by_gas_price(transactions: &mut [VerifiedExecutableTransaction]) {
         let _scope = monitored_scope("ConsensusCommitHandler::order_by_gas_price");
         transactions.sort_by_key(|tx| {
             // Reverse order, so that transactions with higher gas price are put to the beginning.
