+++ crates/sui-core/src/unit_tests/authority_tests.rs
@@ -4,12 +4,12 @@
 
 use bcs;
 use fastcrypto::traits::KeyPair;
-use futures::{stream::FuturesUnordered, StreamExt};
+use futures::{StreamExt, stream::FuturesUnordered};
 use insta::assert_snapshot;
 use move_binary_format::{
+    CompiledModule,
     file_format::{self, AddressIdentifierIndex, IdentifierIndex, ModuleHandle},
     file_format_common::BinaryConstants,
-    CompiledModule,
 };
 use move_core_types::identifier::IdentStr;
 use move_core_types::language_storage::StructTag;
@@ -17,7 +17,7 @@ use move_core_types::{
     account_address::AccountAddress, ident_str, identifier::Identifier, language_storage::TypeTag,
 };
 use rand::seq::SliceRandom;
-use rand::{prelude::StdRng, SeedableRng};
+use rand::{SeedableRng, prelude::StdRng};
 use serde_json::json;
 use std::collections::HashSet;
 use std::fs;
@@ -50,14 +50,14 @@ use sui_types::utils::{
     to_sender_signed_transaction, to_sender_signed_transaction_with_multi_signers,
 };
 use sui_types::{
-    base_types::{dbg_addr, FullObjectRef},
-    crypto::{get_key_pair, Signature},
-    crypto::{AccountKeyPair, AuthorityKeyPair},
-    object::{Owner, GAS_VALUE_FOR_TESTING, OBJECT_START_VERSION},
     MOVE_STDLIB_PACKAGE_ID, SUI_AUTHENTICATOR_STATE_OBJECT_ID, SUI_CLOCK_OBJECT_ID,
     SUI_FRAMEWORK_PACKAGE_ID, SUI_RANDOMNESS_STATE_OBJECT_ID, SUI_SYSTEM_STATE_OBJECT_ID,
+    base_types::{FullObjectRef, dbg_addr},
+    crypto::{AccountKeyPair, AuthorityKeyPair},
+    crypto::{Signature, get_key_pair},
+    object::{GAS_VALUE_FOR_TESTING, OBJECT_START_VERSION, Owner},
 };
-use sui_types::{digests::Digest, SUI_CLOCK_OBJECT_SHARED_VERSION};
+use sui_types::{SUI_CLOCK_OBJECT_SHARED_VERSION, digests::Digest};
 use sui_types::{dynamic_field::DynamicFieldType, messages_consensus::ConsensusTransaction};
 
 use crate::authority::test_authority_builder::TestAuthorityBuilder;
@@ -169,7 +169,7 @@ async fn construct_shared_object_transaction_with_sequence_number(
         .await
         .unwrap();
         effects.status().unwrap();
-        let shared_object_id = effects.created()[0].0 .0;
+        let shared_object_id = effects.created()[0].0.0;
         let mut shared_object = authority.get_object(&shared_object_id).await.unwrap();
         if let Some(initial_shared_version) = initial_shared_version_override {
             shared_object
@@ -363,7 +363,7 @@ async fn test_dev_inspect_object_by_bytes() {
     )
     .await
     .unwrap();
-    let created_object_id = effects.created()[0].0 .0;
+    let created_object_id = effects.created()[0].0.0;
     let created_object = validator.get_object(&created_object_id).await.unwrap();
     let created_object_bytes = created_object
         .data
@@ -466,7 +466,7 @@ async fn test_dev_inspect_unowned_object() {
     )
     .await
     .unwrap();
-    let created_object_id = effects.created()[0].0 .0;
+    let created_object_id = effects.created()[0].0.0;
     let created_object = validator.get_object(&created_object_id).await.unwrap();
     assert!(alice != bob);
     assert_eq!(created_object.owner, Owner::AddressOwner(bob));
@@ -534,7 +534,7 @@ async fn test_dev_inspect_dynamic_field() {
                 .await
                 .unwrap();
                 assert!(effects.status().is_ok(), "{:#?}", effects.status());
-                let created_object_id = effects.created()[0].0 .0;
+                let created_object_id = effects.created()[0].0.0;
                 let created_object = validator.get_object(&created_object_id).await.unwrap();
                 created_object
                     .data
@@ -640,7 +640,7 @@ async fn test_dev_inspect_return_values() {
     )
     .await
     .unwrap();
-    let created_object_id = effects.created()[0].0 .0;
+    let created_object_id = effects.created()[0].0.0;
     let created_object = validator.get_object(&created_object_id).await.unwrap();
     let created_object_bytes = created_object
         .data
@@ -1206,39 +1206,43 @@ async fn test_handle_transfer_transaction_bad_signature() {
     let mut bad_signature_transfer_transaction = transfer_transaction.clone().into_inner();
     *bad_signature_transfer_transaction
         .data_mut_for_testing()
-        .tx_signatures_mut_for_testing() =
-        vec![
-            Signature::new_secure(transfer_transaction.data().intent_message(), &unknown_key)
-                .into(),
-        ];
+        .tx_signatures_mut_for_testing() = vec![
+        Signature::new_secure(transfer_transaction.data().intent_message(), &unknown_key).into(),
+    ];
 
-    assert!(client
-        .handle_transaction(bad_signature_transfer_transaction, None)
-        .await
-        .is_err());
+    assert!(
+        client
+            .handle_transaction(bad_signature_transfer_transaction, None)
+            .await
+            .is_err()
+    );
 
     // This metric does not increment because of the early check for correct sender address in
     // verify_user_input (transaction.rs)
     // assert_eq!(metrics.signature_errors.get(), 1);
 
     let object = authority_state.get_object(&object_id).await.unwrap();
-    assert!(authority_state
-        .get_transaction_lock(
-            &object.compute_object_reference(),
-            &authority_state.epoch_store_for_testing()
-        )
-        .await
-        .unwrap()
-        .is_none());
+    assert!(
+        authority_state
+            .get_transaction_lock(
+                &object.compute_object_reference(),
+                &authority_state.epoch_store_for_testing()
+            )
+            .await
+            .unwrap()
+            .is_none()
+    );
 
-    assert!(authority_state
-        .get_transaction_lock(
-            &object.compute_object_reference(),
-            &authority_state.epoch_store_for_testing()
-        )
-        .await
-        .unwrap()
-        .is_none());
+    assert!(
+        authority_state
+            .get_transaction_lock(
+                &object.compute_object_reference(),
+                &authority_state.epoch_store_for_testing()
+            )
+            .await
+            .unwrap()
+            .is_none()
+    );
 }
 
 #[tokio::test]
@@ -1317,29 +1321,35 @@ async fn test_handle_transfer_transaction_unknown_sender() {
         rgp,
     );
 
-    assert!(authority_state
-        .handle_transaction(&epoch_store, unknown_sender_transfer_transaction)
-        .await
-        .is_err());
+    assert!(
+        authority_state
+            .handle_transaction(&epoch_store, unknown_sender_transfer_transaction)
+            .await
+            .is_err()
+    );
 
     let object = authority_state.get_object(&object_id).await.unwrap();
-    assert!(authority_state
-        .get_transaction_lock(
-            &object.compute_object_reference(),
-            &authority_state.epoch_store_for_testing()
-        )
-        .await
-        .unwrap()
-        .is_none());
+    assert!(
+        authority_state
+            .get_transaction_lock(
+                &object.compute_object_reference(),
+                &authority_state.epoch_store_for_testing()
+            )
+            .await
+            .unwrap()
+            .is_none()
+    );
 
-    assert!(authority_state
-        .get_transaction_lock(
-            &object.compute_object_reference(),
-            &authority_state.epoch_store_for_testing()
-        )
-        .await
-        .unwrap()
-        .is_none());
+    assert!(
+        authority_state
+            .get_transaction_lock(
+                &object.compute_object_reference(),
+                &authority_state.epoch_store_for_testing()
+            )
+            .await
+            .unwrap()
+            .is_none()
+    );
 }
 
 #[tokio::test]
@@ -1375,21 +1385,25 @@ async fn test_handle_transfer_transaction_ok() {
     );
 
     // Check the initial state of the locks
-    assert!(authority_state
-        .get_transaction_lock(
-            &(object_id, before_object_version, object.digest()),
-            &authority_state.epoch_store_for_testing()
-        )
-        .await
-        .unwrap()
-        .is_none());
-    assert!(authority_state
-        .get_transaction_lock(
-            &(object_id, after_object_version, object.digest()),
-            &authority_state.epoch_store_for_testing()
-        )
-        .await
-        .is_err());
+    assert!(
+        authority_state
+            .get_transaction_lock(
+                &(object_id, before_object_version, object.digest()),
+                &authority_state.epoch_store_for_testing()
+            )
+            .await
+            .unwrap()
+            .is_none()
+    );
+    assert!(
+        authority_state
+            .get_transaction_lock(
+                &(object_id, after_object_version, object.digest()),
+                &authority_state.epoch_store_for_testing()
+            )
+            .await
+            .is_err()
+    );
 
     let account_info = authority_state
         .handle_transaction(&epoch_store, transfer_transaction.clone())
@@ -2039,7 +2053,7 @@ async fn test_handle_move_transaction() {
     assert_eq!(effects.created().len(), 1);
     assert_eq!(effects.mutated().len(), 1);
 
-    let created_object_id = effects.created()[0].0 .0;
+    let created_object_id = effects.created()[0].0.0;
     // check that transaction actually created an object with the expected ID, owner
     let created_obj = authority_state
         .get_object(&created_object_id)
@@ -2432,21 +2446,25 @@ async fn test_handle_confirmation_transaction_ok() {
     assert_eq!(next_sequence_number, new_account.version());
 
     // Check locks are set and archived correctly
-    assert!(authority_state
-        .get_transaction_lock(
-            &(object_id, 1.into(), old_account.digest()),
-            &authority_state.epoch_store_for_testing()
-        )
-        .await
-        .is_err());
-    assert!(authority_state
-        .get_transaction_lock(
-            &(object_id, 2.into(), new_account.digest()),
-            &authority_state.epoch_store_for_testing()
-        )
-        .await
-        .expect("Exists")
-        .is_none());
+    assert!(
+        authority_state
+            .get_transaction_lock(
+                &(object_id, 1.into(), old_account.digest()),
+                &authority_state.epoch_store_for_testing()
+            )
+            .await
+            .is_err()
+    );
+    assert!(
+        authority_state
+            .get_transaction_lock(
+                &(object_id, 2.into(), new_account.digest()),
+                &authority_state.epoch_store_for_testing()
+            )
+            .await
+            .expect("Exists")
+            .is_none()
+    );
 }
 
 #[tokio::test]
@@ -2748,10 +2766,12 @@ async fn test_move_call_delete() {
 async fn test_get_latest_parent_entry_genesis() {
     let authority_state = TestAuthorityBuilder::new().build().await;
     // There should not be any object with ID zero
-    assert!(authority_state
-        .get_object_or_tombstone(ObjectID::ZERO)
-        .await
-        .is_none());
+    assert!(
+        authority_state
+            .get_object_or_tombstone(ObjectID::ZERO)
+            .await
+            .is_none()
+    );
 }
 
 #[tokio::test]
@@ -2783,7 +2803,7 @@ async fn test_get_latest_parent_entry() {
     .unwrap();
     let (new_object_id2, seq2, _) = effects.created()[0].0;
 
-    let update_version = SequenceNumber::lamport_increment([seq1, seq2, effects.gas_object().0 .1]);
+    let update_version = SequenceNumber::lamport_increment([seq1, seq2, effects.gas_object().0.1]);
 
     let effects = call_move(
         &authority_state,
@@ -2810,7 +2830,7 @@ async fn test_get_latest_parent_entry() {
     assert_eq!(obj_ref.0, new_object_id1);
     assert_eq!(obj_ref.1, update_version);
 
-    let delete_version = SequenceNumber::lamport_increment([obj_ref.1, effects.gas_object().0 .1]);
+    let delete_version = SequenceNumber::lamport_increment([obj_ref.1, effects.gas_object().0.1]);
 
     let _effects = call_move(
         &authority_state,
@@ -2834,10 +2854,12 @@ async fn test_get_latest_parent_entry() {
     // Prevent overflow
     x[last_index] = u8::MAX - x[last_index];
     let unknown_object_id: ObjectID = x.try_into().unwrap();
-    assert!(authority_state
-        .get_object_or_tombstone(unknown_object_id)
-        .await
-        .is_none());
+    assert!(
+        authority_state
+            .get_object_or_tombstone(unknown_object_id)
+            .await
+            .is_none()
+    );
 
     // Check gas object is returned.
     let obj_ref = authority_state
@@ -3277,7 +3299,7 @@ async fn test_transfer_sui_no_amount() {
     // and got transferred. Also check on its version and new balance.
     assert!(effects.status().is_ok());
     assert!(effects.mutated_excluding_gas().is_empty());
-    assert!(gas_ref.1 < effects.gas_object().0 .1);
+    assert!(gas_ref.1 < effects.gas_object().0.1);
     assert_eq!(effects.gas_object().1, Owner::AddressOwner(recipient));
     let new_balance =
         sui_types::gas::get_gas_balance(&authority_state.get_object(&gas_object_id).await.unwrap())
@@ -3320,11 +3342,11 @@ async fn test_transfer_sui_with_amount() {
     assert_eq!(effects.created().len(), 1);
     assert_eq!(effects.created()[0].1, Owner::AddressOwner(recipient));
     let new_gas = authority_state
-        .get_object(&effects.created()[0].0 .0)
+        .get_object(&effects.created()[0].0.0)
         .await
         .unwrap();
     assert_eq!(sui_types::gas::get_gas_balance(&new_gas).unwrap(), 500);
-    assert!(gas_ref.1 < effects.gas_object().0 .1);
+    assert!(gas_ref.1 < effects.gas_object().0.1);
     assert_eq!(effects.gas_object().1, Owner::AddressOwner(sender));
     let new_balance =
         sui_types::gas::get_gas_balance(&authority_state.get_object(&gas_object_id).await.unwrap())
@@ -3464,7 +3486,7 @@ async fn test_clear_cache_reverts_wrap_move_call() {
 
     // The gas is uncharged
     let gas = cache.get_object(&gas_object_id).unwrap();
-    assert_eq!(gas.version(), create_effects.gas_object().0 .1);
+    assert_eq!(gas.version(), create_effects.gas_object().0.1);
 }
 
 #[tokio::test]
@@ -3544,7 +3566,7 @@ async fn test_clear_cache_reverts_unwrap_move_call() {
     assert_eq!(unwrap_effects.deleted().len(), 1);
     assert_eq!(unwrap_effects.deleted()[0].0, wrapper_v0.0);
     assert_eq!(unwrap_effects.unwrapped().len(), 1);
-    assert_eq!(unwrap_effects.unwrapped()[0].0 .0, object_v0.0);
+    assert_eq!(unwrap_effects.unwrapped()[0].0.0, object_v0.0);
 
     let cache = &authority_state.get_object_cache_reader();
     let reconfig_api = authority_state.get_reconfig_api();
@@ -3561,7 +3583,7 @@ async fn test_clear_cache_reverts_unwrap_move_call() {
 
     // The gas is uncharged
     let gas = cache.get_object(&gas_object_id).unwrap();
-    assert_eq!(gas.version(), wrap_effects.gas_object().0 .1);
+    assert_eq!(gas.version(), wrap_effects.gas_object().0.1);
 }
 
 #[tokio::test]
@@ -5069,16 +5091,29 @@ async fn test_consensus_commit_prologue_generation() {
             .unwrap(),
     );
 
-    let (processed_consensus_transactions, assigned_versions) =
-        send_batch_consensus_no_execution(&authority_state, &certificates, false).await;
+    // Set up ConsensusHandler for testing
+    let consensus_setup =
+        crate::consensus_test_utils::setup_consensus_handler_for_testing(&authority_state).await;
+    let mut consensus_handler = consensus_setup.consensus_handler;
+    let captured_transactions = consensus_setup.captured_transactions;
+
+    let (processed_consensus_transactions, assigned_versions) = send_batch_consensus_no_execution(
+        &authority_state,
+        &certificates,
+        &mut consensus_handler,
+        &captured_transactions,
+    )
+    .await;
 
     let assigned_versions = assigned_versions.into_map();
 
     // Consensus commit prologue V2 should be turned on everywhere.
-    assert!(authority_state
-        .epoch_store_for_testing()
-        .protocol_config()
-        .include_consensus_digest_in_prologue());
+    assert!(
+        authority_state
+            .epoch_store_for_testing()
+            .protocol_config()
+            .include_consensus_digest_in_prologue()
+    );
 
     // Tests that new consensus commit prologue transaction is added to the batch, and it is the first transaction.
     assert_eq!(processed_consensus_transactions.len(), 3);
@@ -5112,7 +5147,7 @@ async fn test_consensus_commit_prologue_generation() {
     };
     // Compare clock version between the prologue (index 0) and the transaction that uses clock (index 2)
     let clock_v1 = get_assigned_version(&processed_consensus_transactions[0].key());
-    let clock_v2 = get_assigned_version(&processed_consensus_transactions[2].key());
+    let clock_v2 = get_assigned_version(&processed_consensus_transactions[1].key());
     assert!(clock_v1 < clock_v2);
 }
 
@@ -5465,7 +5500,7 @@ async fn test_gas_smashing() {
             assert!(effects.status().is_err());
         }
         // gas object in effects is first coin in vector of coins
-        assert_eq!(gas_coin_ids[0], effects.gas_object().0 .0);
+        assert_eq!(gas_coin_ids[0], effects.gas_object().0.0);
         // object is created on success and gas at position 0 mutated
         let created = usize::from(success);
         assert_eq!(
@@ -5475,10 +5510,12 @@ async fn test_gas_smashing() {
         // extra coin are deleted
         assert_eq!(effects.deleted().len() as u64, coin_num - 1);
         for gas_coin_id in &gas_coin_ids[1..] {
-            assert!(effects
-                .deleted()
-                .iter()
-                .any(|deleted| deleted.0 == *gas_coin_id));
+            assert!(
+                effects
+                    .deleted()
+                    .iter()
+                    .any(|deleted| deleted.0 == *gas_coin_id)
+            );
         }
         // balance on first coin is correct
         let balance =
@@ -6310,10 +6347,12 @@ async fn test_consensus_handler_per_object_congestion_control(
     .await;
     // The last deferred transaction should go through
     assert_eq!(scheduled_txns.0.len(), 1);
-    assert!(authority
-        .epoch_store_for_testing()
-        .get_all_deferred_transactions_for_test()
-        .is_empty());
+    assert!(
+        authority
+            .epoch_store_for_testing()
+            .get_all_deferred_transactions_for_test()
+            .is_empty()
+    );
 }
 
 // Helper to process certificates through consensus handler directly
@@ -6555,7 +6594,7 @@ async fn test_consensus_handler_congestion_control_transaction_cancellation() {
     )
     .await;
     assert_eq!(scheduled_txns.len(), 2); // 1 user transaction + 1 prologue
-                                         // Note that consensus handler also generates consensus commit prologue transaction, and it must be the first one.
+    // Note that consensus handler also generates consensus commit prologue transaction, and it must be the first one.
     assert!(matches!(
         scheduled_txns[0]
             .as_tx()
@@ -6614,10 +6653,12 @@ async fn test_consensus_handler_congestion_control_transaction_cancellation() {
         )
         .await;
     assert_eq!(scheduled_txns.len(), 3); // 3 = 2 user transactions + 1 consensus commit prologue transaction.
-    assert!(authority
-        .epoch_store_for_testing()
-        .get_all_deferred_transactions_for_test()
-        .is_empty());
+    assert!(
+        authority
+            .epoch_store_for_testing()
+            .get_all_deferred_transactions_for_test()
+            .is_empty()
+    );
 
     // Check cancelled transaction shared locks.
     let assigned_versions = assigned_versions.into_map();
